# server/swagger.yml
openapi: 3.1.0
info:
  title: PLANT System API
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000
    description: 本機開發環境
security: []
...
paths:
  /api/bootstrap:
    get:
      tags: [Bootstrap]
      summary: 取得前端初始化資料
      operationId: getBootstrap
      security:
        - ActivationCookie: []
        - {}
      responses:
        "200":
          description: 初始化資料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/projects:
    post:
      tags: [Projects]
      summary: 建立專案
      operationId: createProject
      security:
        - ActivationCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreate"
            example:
              rough_idea: 建立一個訂餐外送需求開發專案
      responses:
        "200":
          description: 專案建立成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/projects/{project_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Projects]
      summary: 取得專案 Artifact
      operationId: getProject
      security:
        - ActivationCookie: []
        - {}
      responses:
        "200":
          description: 專案資料
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Projects]
      summary: 刪除專案
      operationId: deleteProject
      security:
        - ActivationCookie: []
      responses:
        "200":
          description: 專案刪除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteProjectResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 專案正在執行 Run，或有附件、刪除等專案操作正在進行，無法刪除
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                detail: Project is running or another project operation is in progress
  /api/projects/{project_id}/cost-summary:
    get:
      tags: [Projects]
      summary: 取得專案成本摘要
      operationId: getProjectCostSummary
      security:
        - ActivationCookie: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: 成本摘要
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CostSummaryResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/projects/{project_id}/references:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [References]
      summary: 列出專案參考文件
      operationId: listProjectReferences
      security:
        - ActivationCookie: []
        - {}
      responses:
        "200":
          description: 參考文件列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceListResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [References]
      summary: 上傳專案參考文件
      operationId: uploadProjectReference
      security:
        - ActivationCookie: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 上傳成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 同名附件已存在、專案正在執行 Run，或有其他專案操作正在進行
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              examples:
                duplicateReference:
                  summary: 同名附件已存在
                  value:
                    detail: "Reference already exists: 規定.docx"
                projectBusy:
                  summary: 專案正在執行或有其他操作
                  value:
                    detail: Project is running or another project operation is in progress
        "413":
          description: 附件超過 25 MiB 上限
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                detail: File exceeds 26214400 byte limit
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/projects/{project_id}/references/{name}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/ReferenceName"
    get:
      tags: [References]
      summary: 下載或預覽參考文件
      operationId: downloadProjectReference
      security:
        - ActivationCookie: []
        - {}
      parameters:
        - $ref: "#/components/parameters/Inline"
      responses:
        "200":
          $ref: "#/components/responses/BinaryFile"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [References]
      summary: 刪除參考文件
      operationId: deleteProjectReference
      security:
        - ActivationCookie: []
      responses:
        "200":
          description: 刪除成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceMutationResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: 專案正在執行 Run，或有其他專案操作正在進行，無法刪除附件
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                detail: Project is running or another project operation is in progress
  /{project_id}/references/{name}:
    get:
      tags: [References]
      summary: 透過公開路徑下載或預覽參考文件
      operationId: publicDownloadProjectReference
      security:
        - ActivationCookie: []
        - {}
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ReferenceName"
        - $ref: "#/components/parameters/Inline"
      responses:
        "200":
          $ref: "#/components/responses/BinaryFile"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    ActivationCookie:
      type: apiKey
      in: cookie
      name: plant_activation
      description: 透過 POST /api/secrets/activation-code 取得的 HttpOnly Cookie
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema: { type: string }
      example: "173949223214"
...
    ReferenceName:
      name: name
      in: path
      required: true
      schema: { type: string }
      example: 規定.docx
...
    Inline:
      name: inline
      in: query
      required: false
      schema: { type: boolean, default: false }
      description: true 時以 inline Content-Disposition 回傳
  responses:
    BadRequest:
      description: 請求資料或操作條件不合法
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: 缺少啟動權限或專案目前不可公開讀取
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example: { detail: 需要啟動碼才能執行此操作 }
    NotFound:
      description: 找不到指定資源
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ValidationError:
      description: FastAPI/Pydantic 請求格式驗證失敗
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
    InternalServerError:
      description: 伺服器內部錯誤
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    BinaryFile:
      description: 檔案內容；實際 Content-Type 依副檔名決定
      content:
        application/octet-stream:
          schema: { type: string, format: binary }
...
  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          description: 錯誤訊息或詳細資料
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            additionalProperties: true
    ProjectCreate:
      type: object
      required: [rough_idea]
      properties:
        rough_idea:
          type: string
          minLength: 1
    ProjectCreated:
      type: object
      required: [project_id, rough_idea]
      properties:
        project_id: { type: string }
        rough_idea: { type: string }
    ProjectSummary:
      type: object
      required: [project_id]
      properties:
        project_id: { type: string }
        created_at: { type: [string, "null"] }
        rough_idea: { type: [string, "null"] }
        scenario: { type: [string, "null"] }
        has_results: { type: boolean }
        has_cost_summary: { type: boolean }
        status_hint: { type: [string, "null"] }
        active_run:
          oneOf:
            - $ref: "#/components/schemas/ActiveRunSummary"
            - type: "null"
      additionalProperties: true
    ProjectResponse:
      type: object
      required: [project_id, project]
      properties:
        project_id: { type: string }
        project:
          type: object
          additionalProperties: true
    DeleteProjectResponse:
      type: object
      required: [project_id, deleted]
      properties:
        project_id: { type: string }
        deleted: { type: boolean, const: true }
    BootstrapResponse:
      type: object
      required:
        - config
        - model_summary
        - projects
        - active_runs
        - interrupted_run_count
        - formal_meeting_enabled
        - requires_rounds_input
      properties:
        config:
          $ref: "#/components/schemas/LoadStatus"
        model_summary: { type: string }
        activated: { type: boolean }
        projects:
          type: array
          items: { $ref: "#/components/schemas/ProjectSummary" }
        active_runs:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ActiveRunSummary"
        interrupted_run_count: { type: integer, minimum: 0 }
        formal_meeting_enabled: { type: boolean }
        requires_rounds_input: { type: boolean }
    LoadStatus:
      type: object
      required: [loaded]
      properties:
        loaded: { type: boolean }
        error: { type: [string, "null"] }
    ReferenceInfo:
      type: object
      required: [name, size]
      properties:
        name: { type: string }
        size: { type: integer, minimum: 0 }
    ReferenceListResponse:
      type: object
      required: [project_id, references]
      properties:
        project_id: { type: string }
        references:
          type: array
          items: { $ref: "#/components/schemas/ReferenceInfo" }
    ReferenceMutationResponse:
      type: object
      properties:
        project_id: { type: string }
        name: { type: string }
        size: { type: integer }
        saved: { type: boolean }
        deleted: { type: boolean }
      additionalProperties: true
...
    CostSummaryResponse:
      type: object
      required: [project_id, cost_summary]
      properties:
        project_id: { type: string }
        cost_summary:
          type: object
          additionalProperties: true
    RunStatus:
      type: string
      enum:
        - queued
        - running
        - waiting_for_human
        - cancelling
        - completed
        - failed
        - cancelled
        - interrupted
    ActiveRunSummary:
      type: object
      required: [run_id, status]
      properties:
        run_id: { type: string }
        status: { $ref: "#/components/schemas/RunStatus" }
        pending_decision:
          oneOf:
            - $ref: "#/components/schemas/PendingDecision"
            - type: "null"
...
    PendingDecision:
      type: object
      required: [id, kind, title, description]
      properties:
        id: { type: string }
        kind:
          type: string
          enum:
            - stakeholder_selection
            - human_decision
            - stakeholder_statement_review
            - requirements_review
            - domain_research_review
            - scope_review
            - meeting_issue_proposal_review
        title: { type: string }
        description: { type: string }
        proposed:
          type: array
          items:
            type: object
            additionalProperties: true
        max_select: { type: integer }
        issue:
          type: object
          additionalProperties: true
        options:
          type: object
          additionalProperties: true
...